feature/list_cmd: update dependency , refactor code
