+++ go.mod
@@ -3,12 +3,13 @@ module github.com/cyka/kubectl-java
 go 1.15
 
 require (
-	github.com/go-openapi/strfmt v0.20.0 // indirect
-	github.com/jedib0t/go-pretty v4.3.0+incompatible
-	github.com/mattn/go-runewidth v0.0.10 // indirect
+	github.com/fatih/color v1.7.0
+	github.com/kr/text v0.2.0 // indirect
+	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/spf13/cobra v1.1.1
 	github.com/spf13/pflag v1.0.5
-	k8s.io/apimachinery v0.20.2
+	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200605160147-a5ece683394c // indirect
 	k8s.io/cli-runtime v0.20.2
 	k8s.io/client-go v0.20.2
 )
